requirements.txt
chardet==3.0.4
chardet2==2.0.3
flask-reuploaded==1.3.0
flask-uploads==0.2.1
google-api-python-client==2.81.0
google-auth-oauthlib==1.0.0
ipykernel==6.21.3
librosa==0.10.0
oauth2==1.9.0.post1
pillow==9.4.0
pip-chill==1.0.1
pyvimeo==1.1.0
simplejson==3.18.3
